+++ isis/src/base/apps/crop/main.cpp
@@ -2,9 +2,10 @@
 
 #include "crop.h"
 
-using namespace std; 
+using namespace std;
 using namespace Isis;
 
+// logic is in crop.cpp
 void IsisMain() {
   UserInterface &ui = Application::GetUserInterface();
   PvlGroup results = crop(ui);